modify some logics about constructing model
